+++ src/screens/FleetScreen.js
@@ -1,7 +1,14 @@
 import { useNavigation } from "@react-navigation/native";
 import React, { useContext } from "react";
-import { SafeAreaView, ScrollView, StyleSheet, View } from "react-native";
+import {
+  Alert,
+  SafeAreaView,
+  ScrollView,
+  StyleSheet,
+  View,
+} from "react-native";
 import { IconButton, Text } from "react-native-paper";
+import { deleteActivity } from "../../services/activityServices";
 import { deleteVehicle } from "../../services/fleetServices";
 import Vehicle from "../components/Vehicle";
 import { AppContext } from "../contexts/AppContext";
@@ -12,27 +19,57 @@ const FleetScreen = () => {
   const [state, setState] = useContext(AppContext);
   const { showNotification } = useContext(NotificationContext);
 
-  const handleDelete = async (vehicle) => {
-    try {
-      setState({ ...state, loading: true });
+  const handleDelete = (vehicle) => {
+    Alert.alert(
+      "Delete Confirmation",
+      `Are you sure you want to delete ${vehicle.name}? This will also delete all associated activities.`,
+      [
+        {
+          text: "Cancel",
+        },
+        {
+          text: "Confirm",
+          onPress: async () => {
+            try {
+              setState({ ...state, loading: true });
 
-      await deleteVehicle(vehicle.id);
+              await deleteVehicle(vehicle.id);
 
-      const newFleet = [...state.fleet].filter((v) => v.id !== vehicle.id);
+              const activitesForDelete = state.activities.filter(
+                (a) => a.vehicleId === vehicle.id
+              );
 
-      setState({
-        ...state,
-        fleet: newFleet,
-        loading: false,
-      });
-      showNotification(
-        "success",
-        `${vehicle.name} successfully removed from your fleet.`
-      );
-    } catch (error) {
-      showNotification("error", "Failed to remove vehicle. Please try again.");
-      setState({ ...state, loading: false });
-    }
+              await Promise.all(
+                activitesForDelete.map((a) => deleteActivity(a.id))
+              );
+
+              const newFleet = state.fleet.filter((v) => v.id !== vehicle.id);
+
+              const newActivities = state.activities.filter(
+                (a) => a.vehicleId !== vehicle.id
+              );
+
+              setState({
+                ...state,
+                fleet: newFleet,
+                activities: newActivities,
+                loading: false,
+              });
+              showNotification(
+                "success",
+                `${vehicle.name} successfully removed from your fleet.`
+              );
+            } catch (error) {
+              showNotification(
+                "error",
+                "Failed to remove vehicle. Please try again."
+              );
+              setState({ ...state, loading: false });
+            }
+          },
+        },
+      ]
+    );
   };
 
   const handleEdit = (vehicle) => {
